scraper/requirements.txt
psycopg2_binary==2.9.10
selenium==4.29.0
chromedriver-autoinstaller